r355153 introduced a build failure on a build bot that uses clang natively
on an armv7-a machine. This a temporary fix to use size_t rather than
uint64_t.

llvm-svn: 355195

+++ lld/ELF/Writer.cpp
@@ -85,7 +85,7 @@ private:
 
   std::vector<PhdrEntry *> Phdrs;
 
-  uint64_t FileSize;
+  size_t FileSize;
   uint64_t SectionHeaderOff;
 };
 } // anonymous namespace
